feat: NOP logger if nil ptr to logger
