+++ pkg/logging/utils_test.go
@@ -27,3 +27,27 @@ func TestChildLogger(t *testing.T) {
 	require.Contains(t, msg, "service=child")
 	require.Contains(t, msg, `msg="debug message"`)
 }
+
+func TestSprintf(t *testing.T) {
+	// given:
+	stringWriter := &logging.TestWriter{}
+	logger := logging.New().
+		WithLevel(slog.LevelDebug).
+		WithHandler(logging.TextHandler, stringWriter).
+		Logger()
+
+	// when:
+	logging.Sprintf(logger, slog.LevelInfo, "info message: %d", 123)
+
+	// then:
+	msg := stringWriter.String()
+	require.Contains(t, msg, "info message: 123")
+}
+
+func TestNopIfNil(t *testing.T) {
+	// when:
+	logger := logging.NopIfNil(nil)
+
+	// then:
+	require.NotNil(t, logger)
+}
